Run 'forge fmt'

+++ src/VeganRestaurant.sol
@@ -4,11 +4,11 @@ pragma solidity ^0.8.13;
 contract VeganRestaurant {
     struct Restaurant {
         address owner;
-        uint avgRating;
-        uint totalRatings;
+        uint256 avgRating;
+        uint256 totalRatings;
     }
     // Version number for the contract.
-    uint public version = 1;
+    uint256 public version = 1;
     // Mapping from the Open Location Code (Plus Code) to its details.
     mapping(string => Restaurant) public restaurants;
 
@@ -16,10 +16,6 @@ contract VeganRestaurant {
         require(bytes(plusCode).length > 0, "Plus code cannot be empty.");
         require(restaurants[plusCode].owner == address(0), "Restaurant already exists.");
 
-        restaurants[plusCode] = Restaurant({
-            owner: msg.sender,
-            avgRating: 0,
-            totalRatings: 0
-        });
+        restaurants[plusCode] = Restaurant({owner: msg.sender, avgRating: 0, totalRatings: 0});
     }
 }
